move: test fixture 디렉토리 이동
