+++ backend/src/test/java/com/sjiwon/anotherart/art/infra/query/ArtDetailQueryRepositoryTest.java
@@ -8,10 +8,10 @@ import com.sjiwon.anotherart.art.utils.search.ArtDetailSearchCondition;
 import com.sjiwon.anotherart.auction.domain.Auction;
 import com.sjiwon.anotherart.auction.domain.AuctionRepository;
 import com.sjiwon.anotherart.common.RepositoryTest;
+import com.sjiwon.anotherart.common.fixture.ArtFixture;
+import com.sjiwon.anotherart.common.fixture.MemberFixture;
 import com.sjiwon.anotherart.favorite.domain.Favorite;
 import com.sjiwon.anotherart.favorite.domain.FavoriteRepository;
-import com.sjiwon.anotherart.fixture.ArtFixture;
-import com.sjiwon.anotherart.fixture.MemberFixture;
 import com.sjiwon.anotherart.member.domain.Member;
 import com.sjiwon.anotherart.member.domain.MemberRepository;
 import jakarta.persistence.EntityManager;
@@ -40,8 +40,8 @@ import static com.sjiwon.anotherart.art.utils.search.SortType.LIKE_ASC;
 import static com.sjiwon.anotherart.art.utils.search.SortType.LIKE_DESC;
 import static com.sjiwon.anotherart.art.utils.search.SortType.PRICE_ASC;
 import static com.sjiwon.anotherart.art.utils.search.SortType.PRICE_DESC;
-import static com.sjiwon.anotherart.fixture.AuctionFixture.AUCTION_OPEN_NOW;
-import static com.sjiwon.anotherart.fixture.MemberFixture.MEMBER_A;
+import static com.sjiwon.anotherart.common.fixture.AuctionFixture.AUCTION_OPEN_NOW;
+import static com.sjiwon.anotherart.common.fixture.MemberFixture.MEMBER_A;
 import static com.sjiwon.anotherart.member.domain.point.PointType.CHARGE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
