+++ backend/src/test/java/com/sjiwon/anotherart/global/security/handler/AjaxAuthenticationSuccessHandlerTest.java
@@ -7,9 +7,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
+import static com.sjiwon.anotherart.common.fixture.MemberFixture.MEMBER_A;
 import static com.sjiwon.anotherart.common.utils.TokenUtils.ACCESS_TOKEN;
 import static com.sjiwon.anotherart.common.utils.TokenUtils.REFRESH_TOKEN;
-import static com.sjiwon.anotherart.fixture.MemberFixture.MEMBER_A;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
