+++ backend/src/test/java/com/sjiwon/anotherart/member/controller/MemberInformationApiControllerTest.java
@@ -4,7 +4,7 @@ import com.sjiwon.anotherart.art.infra.query.dto.response.AuctionArt;
 import com.sjiwon.anotherart.art.infra.query.dto.response.BasicArt;
 import com.sjiwon.anotherart.art.infra.query.dto.response.BasicAuction;
 import com.sjiwon.anotherart.common.ControllerTest;
-import com.sjiwon.anotherart.fixture.MemberFixture;
+import com.sjiwon.anotherart.common.fixture.MemberFixture;
 import com.sjiwon.anotherart.member.domain.Member;
 import com.sjiwon.anotherart.member.domain.point.PointType;
 import com.sjiwon.anotherart.member.infra.query.dto.response.BasicMember;
@@ -27,13 +27,13 @@ import java.util.List;
 
 import static com.sjiwon.anotherart.art.domain.ArtStatus.ON_SALE;
 import static com.sjiwon.anotherart.art.domain.ArtStatus.SOLD;
+import static com.sjiwon.anotherart.common.fixture.ArtFixture.AUCTION_1;
+import static com.sjiwon.anotherart.common.fixture.ArtFixture.AUCTION_2;
+import static com.sjiwon.anotherart.common.fixture.ArtFixture.AUCTION_3;
+import static com.sjiwon.anotherart.common.fixture.ArtFixture.GENERAL_1;
+import static com.sjiwon.anotherart.common.fixture.MemberFixture.MEMBER_A;
 import static com.sjiwon.anotherart.common.utils.TokenUtils.ACCESS_TOKEN;
 import static com.sjiwon.anotherart.common.utils.TokenUtils.BEARER_TOKEN;
-import static com.sjiwon.anotherart.fixture.ArtFixture.AUCTION_1;
-import static com.sjiwon.anotherart.fixture.ArtFixture.AUCTION_2;
-import static com.sjiwon.anotherart.fixture.ArtFixture.AUCTION_3;
-import static com.sjiwon.anotherart.fixture.ArtFixture.GENERAL_1;
-import static com.sjiwon.anotherart.fixture.MemberFixture.MEMBER_A;
 import static com.sjiwon.anotherart.member.domain.point.PointType.CHARGE;
 import static com.sjiwon.anotherart.member.domain.point.PointType.PURCHASE;
 import static com.sjiwon.anotherart.member.domain.point.PointType.REFUND;
