+++ backend/src/test/resources/application.yml
@@ -1,16 +1,11 @@
 spring:
-  datasource:
-    url: jdbc:h2:mem:test;MODE=MYSQL;
-    username: sa
-    password:
-    driver-class-name: org.h2.Driver
-
   jpa:
     open-in-view: false
     hibernate:
       ddl-auto: create
     properties:
       hibernate:
+        dialect: org.hibernate.dialect.MySQLDialect
         show_sql: false
         format_sql: true
         default_batch_fetch_size: 50
